# [process][windows] Use win32 API in process.Children() instead of slow WMI call